+++ History.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * "dev" format will use same tokens as other formats
+  * `:response-time` token is now empty when immediate used
   * `:response-time` token is now monotonic
   * `:response-time` token has precision to 1 μs
   * fix `:status` + immediate output in node.js 0.8